funcsをutilsディレクトリに移動
